gcp_vertex/pyproject.toml
[project]
name = "gcp-vertex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.8"
dependencies = [
    "langchain>=1.2.15",
    "langchain-google-genai>=4.2.1",
    "python-dotenv>=1.2.2",
]
